Add grouped links functionality to pagination
This changeset adds a number of options to the paginate method and the
will_paginate view helper to enable the use of grouped links. The grouping
works by creating a page link to the first page that has a result value
matching that group. It requires that you have a database attribute which
can be grouped (or that you create one especially for this feature).

Examples of use:

Display of a list of recipies for fruit. You might want to group all the
recipies for a particular fruit together while displaying the name in
alphabetical order. You should have a 'fruit' database attribute and then
pass that to the paginate method with:

 @recipes = Recipe.paginate :page => params[:page], :group_by => 'fruit', :order => 'fruit ASC, name ASC'

or

Display a list of people. You might want to group them by the first letter
of their name. You would create a new database attribute called 'first_letter',
and update this with a +before_save+ method on your model. Then you can
paginate it with:

 @people = Person.paginate :page => params[:page], :group_by => 'first_letter', :order => 'first_letter ASC, name ASC'