Add functionality to extend links collection with other entries.
Sometimes when using grouped links it is useful to be able to add other entries, even
if there is no page containing those values. For example, you may want a 'first letter'
grouping which shows all the letters, even if only some of them are present. This
changeset adds a +add_missing_links+ method to the collection which takes an array
of strings and adds entries for these if they don't already exist. When rendered, these
entries appear as a span (rather than a link), with a css class of 'missing'.